Remove a candidate measure format
The measure format where each of the microsegment categories (e.g., end use, fuel type) were specified using an object with a boolean list for each possible case of each category. This approach to formatting the measure yielded an exceedingly long measure (in number of lines) and did not make the measure more human-readable. I don't believe it would have made it easier to enter the measure either. I think the 'alt2 microsegement data format' is the best (of the current approaches) way to enable multiple cases for the categories describing the microsegments.