Added test for residential data integrity
The test reads in the header line from the residential AEO micro
segments RESDBOUT.txt file, then compares this line to a Python regular
expression that reflects our expectations about what each column in the
file should be covering.  If there is no match, the header line is
printed for diagnostic purposes.  The regex comparison checks for the
presence of key letters in each column head title (case insensitive).