Microsegment Updating Routine (Residential)
This program reads in market microsegment names from the
microsegments.json file and updates the energy use projections attached
to each microsegment using the most recent EIA Annual Energy Outlook
source information.

The file has been developed initially around the residential sector,
but may be adapted to the commercial sector.  It has been tested on a
reduced version of the microsegments.json file (residential only, for
one census division and all building types, fuels, end uses, and
technologies).  The test runs in about 7 minutes.  Note that the
program does not currently write out an updated json file, but will do
so in the future.  Note also that formal tests for each of the
functions included in this file will be written into the mseg_test.py
file in the coming days.

A few key points about the code:

1.) A series of dicts are created to translate microsegment names in
the microsegments.json file into filtering information used to find
corresponding energy use projections for each microsegment in the AEO
.txt file.

2.) The main function (“value_replacer_main”) first imports the
microsegments.json and AEO .txt files.  The json file is read in as a
dict and the .txt file is read in as a list with each of its rows as
list elements.  The function then runs through every microsegment key
in the dict, finds the set of lines in the .txt containing energy use
information for that microsegment, and stores the energy use
information for those lines as a list.  This list is then used to
update the values attached to each microsegment key in the dict.
Currently, each microsegment list should contain 32 values
(corresponding to 32 years of energy projections in AEO per
microsegment, 2009-2040).

3.)  Key supporting functions for “value_replacer_main” include
“json_translator” (translates microsegment names in the json to row
filtering information for running through the .txt); and
“value_listfinder” (finds the rows in the .txt corresponding to each
microsegment in the json and stores the energy use projections from
those rows as a list).