Updated to almost comply with PEP 8
Revised the formatting of the code to largely comply with indentation guidelines, line length limits, etc. laid out in PEP 8. In the deeply nested for loop, however, and in a few other places, line length limits were relaxed for readability (under the assumption that the code can be reviewed using an editor that offers easy left-right scrolling).