Revised function structure
To help simplify the structure of the code, the value_replacer_listassemble function was eliminated. Now, the value_listfinder function (renamed list_generator) calls the three functions it needs to take the two inputs from the main function and convert them into the values to be output. This structure should be easier to follow at a glance.