Simplified structure of for loop in value_replacer_main
It was discovered that the for loops could run through each level of the JSON file without passing the keys at each level to a separate variable. Eliminating these .keys() calls removed all of the intervening code at the start of the nested for loop in value_replacer_main(). Additionally, the lines where the stock and energy dicts are written to the JSON record were revised to reflect the same simplified structure (referencing msjson directly), rather than relying on the elements of the filterdata variable.