Added new 'txt_parser' function
This function was, for the most part, already included in the code for
the ‘list_generator’ function; this update simply separates out that
bit of code into its own function.

One important addition is the ability for one to specify whether the
routine finds, records, and removes all matched rows of the specified
numpy array (as it has previously), or only finds and removes matched
rows.  This is specified by setting the function’s “command_string”
input to either “Record & Reduce” or to just “Reduce” (case
insensitive).

In “Reduce” only mode, the function may be used to pare down the
initial numpy array created from RESDBOUT.txt.  Here, we would specify
which rows of the initial array are not relevant to the mseg.py
updating routine, and would then use the “txt_parser” function in
“Reduce” mode to remove all those rows from the array before doing
anything further with it.

Note that “txt_parser” in “Record & Reduce” mode is used to find and
record current energy use and stock projections in the RESDBOUT.txt
file, which are ultimately used to update the microsegments json file.