Added new tests for two functions
Two new tests were added, one for the function used to convert the list of fields in the EIA data into regular expressions to be applied to search the EIA data matrix, and one for the function that converts a list longer than 32 (indicated by variable aeo_years) into a list of exactly that length. The former is tested using a set of dummy lists and the corresponding anticipated regex outputs, exploring a range of cases, including demand tuples and lists with tuples inside them. The latter is tested by simply creating a sample numeric list of a random length that is a multiple of 32, reshaping it, and comparing that reshaped list to the output from the function in mseg.py.

Three additional tests were identified but are currently incomplete. These tests will verify the format and structure of the JSON file using the dicts hardcoded in mseg.py, confirm the operation of the list_generator function, and test the input EIA and envelope demand text files to confirm their content has been formatted as byte or unicode strings (as desired).

Additional tests on the new functions created to support importing and processing of the demand data are forthcoming.