Expanded tests on input EIA data file
Added two new tests to the EIA data file to ensure that there are columns named 'consumption' and 'eqstock' are present. Though the headings are capitalized in the code, the regex comparison is not case-sensitive. These tests help to protect the operation of the portion of the txt_parser function in mseg.py that restructures the EIA data.