Updated main function with recursion
The prior main function, 'mseg_updater_main', incorporated deeply nested for loops to traverse the JSON microsegments file and set up a list of keys to be converted into a string to ultimately extract data from the underlying supply, demand, and load data files. In these for loops, additional if statements were required to ensure that the list of keys followed a specific format depending on the particular microsegment being considered. This approach, while relatively easy to follow in the code, was cumbersome and inflexible. Updating the function 'json_translator' to enable it to handle lists in any order facilitated the simplification of the main function.

The prior main function has now been replaced by two functions. The new main function, 'main', only imports (and reduces, where applicable) the supply, demand, and thermal load data and calls the 'walk' function. The 'walk' function traverses the JSON file recursively, replacing the need for the for loops in the prior main function. As it traverses the JSON file, the 'walk' function builds the key list required for matching the appropriate imported data to each leaf node.