Working "demand" microsegment updates & climate conversion
This commit adds a function (“clim_converter”) that converts updated
residential microsegments from a census division breakdown to an AIA
climate zone breakdown.  An associated test for the function has also
been added.

Furthermore, the commit makes a few adjustments to the “mseg.py” code
to ensure correct updating of “demand” microsegments in
“list_generator”.  These are adjustments are described in further
detail below.

The “mseg.py” code passes all tests and runs in a little under 12
minutes for a microsegments_test.json file with just one census
division represented.  Spot checks on the output suggest that
microsegments are being updated and translated to a climate zone basis
correctly.

Details on adjustments to “mseg.py”:

In “json_translator”:

- Changed “if key is ‘demand’” to “if key == ‘demand’” and “if
demand_indicator is not 1” to “if demand_indicator != 1” to yield
desired comparisons.

In “txt_parser”:

- Modified this function to accept “Record” and “Reduce” modes (removed
“Record & Reduce” mode, as the thermal loads input array is now not
being reduced with each run - see note about function output below).
- Changed the “compare to” variable in “txt_parser” to be different for
an EIA array parse vs. a thermal loads array parse.  In the latter,
only the first three elements of the given row being parsed need be
included in the comparison, otherwise the regex matches unwanted rows.
These three elements represent the census division, building type, and
end use matching information for that row, respectively.
- Included the “column_indicator” as an input to “txt_parser” such that
the function produces a single matched value instead of a matched row
for a thermal loads array parse.
- The function no longer reduces the thermal load components input
array, as this was creating errors due to identical coding of secondary
and primary heating microsegments in thermal load components array as
‘HT’.

In “list_generator”:

- Changed the if statement to check for “demand” in the “filterdata”
variable (previously was checking “txt_filter”, which does not include
this type of information).
- Modified the “fuel_remove” regex to handle cases where there is more
than one end use code in the given EIA numpy array row (i.e., ‘(SH|OA)’
in the case of secondary heating).
- In the case of secondary heating, the discovered ‘(SH|OA)’ end use
match is now changed to ‘HT’ for the final “comparefrom_tloads”
variable, as the thermal loads array that this variable is compared to
codes both primary and secondary heating as ‘HT’.