List_generator test & associated updates
Summary of changes:

1) Added a test for “list_generator” function in “mseg_test.py”.  The
test is similar to that for “json_translator”, running three types of
filtering inputs to the function (“ok”, “nonsense”, and “fail”) and
checking that it yields the appropriate output and/or error message for
each.

2) Slightly modified “json_translator” and associated test to ensure
that this function yields an error message whenever the filtering
information for a given microsegment level does not match up with any
of the keys in the associated translation dict defined at the beginning
of “mseg.py” (i.e., cdivdict, bldgtypedict, etc.)  Note that a new set
of “nonsense” filters has also been added to the “json_translator”
test.  These filters do not raise an error in the “json_translator”
routine, but will ultimately yield zeros for stock/energy in
“list_generator”.

3) Made minor formatting improvements throughout, particularly geared
towards reducing the length of certain lines in the code.