Added test for txt_parser function
The txt_parser function, at the moment, performs a few different actions. It is used to reduce the size of the input EIA data by removing unneeded rows, and to select and restructure both the thermal load components data and the EIA data into nearly their final form. The new test for the function checks the operation of these major functions.