Added list of years to JSON database
Previously, the JSON microsegments database did not include any indication of the years corresponding to the lists of stock and consumption data. This omission would prevent the extraction of specific years of data from the JSON and plotting of the data or model outputs. A placeholder for the list of the years associated with the data is now in the JSON at the lowest level in the nested dict structure with the 'years' key. The previous contents of the JSON database are indicated by the 'data' key.