Adjusted clim_converter and associated test
The ‘clim_converter’ function has been adjusted to ensure that cdiv ->
climate conversion weights are attached to the correct census division.
 The census division row number used to find the appropriate conversion
weight in ‘Res_Cdiv_Czone_ConvertTable_Final.txt’ is now specified
using the cdiv translator dict.  If the “input_dict” census division
name cannot be found in this translator dict, the function yields an
error.

In conjunction with these changes to “clim_converter”, the
“ClimConverterTest” has been slightly modified.  The test now checks
whether an error is yielded when one of the input dict census division
names is specified incorrectly relative to the cdiv translator dict in
“mseg.py”.  Input and output list values have also been modified such
that the test will not pass if climate conversion weights are not
applied to the correct census division.