Add years to stock and energy data
To ensure that the data and the years they are associated with do not become unintentionally misaligned, the lists of stock and energy consumption data generated for each leaf/terminal node in the microsegments JSON database are now dicts, where each key is the year, and each value the stock or energy data for the year indicated.

Replacing the lists with dicts required changes to the stock_consume_select function to ensure that when keys were repeated, as with microsegments that combine multiple microsegments in the underlying data (such as secondary heating, which combines SH and OA categories), the values were added together. In the list_generator function, the list comprehension used for multiplying the thermal load component modifiers with the energy use data was changed to a dict comprehension, and the length check for the dicts now ignores empty dicts (where empty dicts indicate a nonsense regex of filter parameters was passed to stock_consume_select).

The tests were updated to reflect the change from lists to dicts for the outputs from the stock_consume_selct and list_generator functions.