Added square footage information to "microsegments.json"
Several microsegments do not have number of units information
associated with their technologies in the AEO, either because this
information does not apply to the technology (i.e., wall insulation),
or because the information is not available (i.e., for commercial
sector microsegments).  This presents a challenge when trying to
normalize each microsegment’s energy consumption to a per unit basis
for easy comparison across microsegments/measures.

In these cases, we can normalize the microsegment energy consumption to
a per square foot basis, as square footage data are available for both
the residential and commercial sectors in the AEO, broken down by
census division and building type.

This commit reflects the incorporation of square footage information
into the existing “microsegments.json” hierarchy.  The square footage
node is incorporated on the same level as fuel type, which facilitates
the integration of this information into the existing “mseg.py”
microsegment updating routines.