Update "mseg.py" & tests to accommodate sq. footage info.
This update to the “mseg.py” routine accommodates new square footage
information in “microsegments.json” from the last commit, and updates
“mseg_test.py” accordingly.

Key changes to “mseg.py” include:

- Remove the “SQ” (square footage) code from the initial EIA supply
numpy array filter, ensuring that the rows with square footage
information remain in this filtered array for later use.

- Add the “SQ” square footage code to the end use translator dict (note
that EIA treats the square footage label as an end use in
“RESDBOUT.txt”).

- Slightly adjust “json_translator” to yield an appropriate output for
a case where square footage filtering information is provided.

- Define a new “sqft_select” function that operates very much like the
existing “stock_consume_select” function, but which generates an output
dict that includes square footage data instead of stock and energy
consumption data.

- Add an “elif” statement to the “list_generator” function to reflect
the slightly different handling of a microsegment square footage update
in this function.

Reflective of these changes, “NumpyArrayReductionTest” in
“mseg_test.py” now incorporates square footage information and adds a
small new routine for testing the “sqft_select” function.  Both the
“DatatoListFormatTest” and “JSONTranslatorTest” have similarly been
updated to incorporate square footage information.

It is noted that a few other tweaks to “mseg_test.py” have been made to
correct minor errors in the previous version of this file.  One of the
main corrections is to the test of the fail filters in
“DatatoListFormatTest” and “JSONTranslator” functions.  Here, the “for”
loop running through the fail filters has been moved ahead of the “with
self.assertRaises(KeyError):” statement; this ensures that all fail
filters are run through in checking for the generation of KeyErrors.

Finally, a few of the fail filters themselves also needed to be
adjusted.  For example, “boiler (electric)” had been included as the
failing element of one filter list, even though this is a valid
technology-level filter name. The name was changed to “boiler (wood
fired),” which generates a KeyError as expected when run through
“json_translator”.