Changed handling of "other fuel" 2nd heating to be consistent
In the case of the “other fuel” fuel type, there are a number of
secondary heating sub-types (i.e. “secondary heating (coal),”
“secondary heating (wood),” etc.)  EIA handles these sub-types on the
technology level, yet this was not the case in our “microsegments.json”
file, leading to possible errors/omissions when parsing secondary
heating information for “other fuel” in the “mseg.py” routine.

The “microsegments.json” has accordingly been modified such that the
secondary heating sub-types for an “other fuel” case are handled on the
technology level under the “supply” node (consistent with how the
multiple heating and cooling technologies are handled across all fuel
types).

Additionally, filtering lists representing both “demand” and “supply”
microsegments for the “other fuel” secondary heating case are now
included in both the “DataToListFormat” and “JSONTranslatorTest”
functions in “mseg.py”.