Added test for "partition_microsegment" function
This commit adds a test to “run_test.py” for the
“partition_microsegment” function used within the “mseg_find_partition”
function of “run.py”.  The “partition_microsegment” function partitions
an input microsegment for a measure to find "competed" stock and energy
consumption as well as "efficient" energy consumption (representing
consumption under the measure).  The function also finds the total cost
of the competed stock for baseline and measure cost levels.

An associated change was made to the “partition_microsegment” function
itself to ensure that the function yields a list of blank elements in a
non “Technical potential” adoption scenario (only the “Technical
potential” adoption scenario is being handled meaningfully by this
function at the moment)