Made handling of square footage stock situations more general
To avoid double counting of stock and associated cost information in
cases where microsegment stock is a square footage, the previous code
had divided master stock and cost values by a reduction factor that
equaled (# end uses contributing to master microsegment * #
technologies contributing to master microsegment).

While this formulation passed all the tests, it was not general enough.
 For example, the reduction factor calculation would not have
accurately handled a case where multiple fuel types contribute to a
measure master microsegment (a case that is not currently tested in
“run_test.py”).

This commit adjusts the reduction factor calculation to ensure that
stock and cost information is effectively only summed on the climate
zone and building type levels when square footage is used as
microsegment stock (as square footage is only broken down on these two
levels).  To do this, the code tracks the number of valid contributing
microsegment key chains for a measure’s master microsegment; this
number is then divided by (# czones the measure covers * # bldg types
the measure covers).