Enable measure performance and cost distributions
“Run.py” now supports the definition of input distributions for measure
performance and cost.  The distributions are sampled with
“numpy.random,” which was chosen over “scipy.stats” because it is
generally considered to run a little faster and because setting up each
sampling call is a bit more straightforward.

“Numpy.random” calls are executed inside a new “rand_list_gen” function
that translates information from the measure definition into the
appropriate format for the “numpy.random” call.  The function currently
supports the normal, lognormal, uniform, triangular, weibull, and gamma
distributions.  If an unsupported distribution name is specified, or if
the parameter specification for a supported distribution is incorrect
(i.e., too few/many parameters), the function yields a ValueError.

Note that the flexibility to define microsegment-specific performance
and cost remains with this commit (even if each is a set of values
rather than a point value as before).  Moreover, the routine allows
flexibility in defining distributions, where one can put a distribution
on cost but not performance, or can put a distribution on the
performance level broken out for climate zone A but not for climate
zone B, if needed.  This will be helpful in sensitivity analyses where,
for example, one may wish to place a distribution on the performance
input but keep the cost input fixed at a single value.

“Run_test.py” has been updated to reflect the new “run.py” distribution
sampling capabilities.  First, a selection of the “ok_measures” in
“FindPartitionMasterMicrosegmentTest” have been copied and assigned
performance/cost input distributions; the cost and performance outputs
yielded by running this new “ok_measures_dist” set are tested in a
“test_mseg_ok_distrib” function.  Output tests include checking the
mean* of output values against a “correct” mean precomputed with a
fixed random seed; and a check that the length of the list of output
values matches the default number of samples specified at the beginning
of “run.py”.

Second, a “RandomSampleTest” has been added to test the “rand_list_gen”
function.  The test runs input distribution information through
“rand_list_gen” to generate a list of sampled output values; it then
uses “scipy.stats” to fit parameter values to the sample using the
appropriate distribution, and checks to see that the sampled parameter
values match precomputed parameter values (again using fixed random
seed).  A separate test checks that when input distribution information
is incorrectly specified, the function yields a ValueError.

Note that the scipy fitting routine is somewhat computationally
expensive, adding ~ 0.15 s to an execution of “run_test.py”, which
otherwise takes ~0.03 s.  Going forward we might delete this portion of
the test or else replace it with something faster.

* The mean is the tested parameter because only normal and lognormal
distributions are used currently in “ok_measures_dist” measures.