Small updates to market calculator appearance
Updated the explanatory text to reference the 2014 Annual Energy Outlook (AEO) instead of the 'latest' AEO, since the 2015 AEO has been released but not yet applied here due to issues with mseg.py not yet addressed. Also updated the classes applied to the 'Recall' and 'Save' buttons so that they are disabled, since they aren't currently able to do anything.