Removed dependence on most objects id selector
To help simplify, and hopefully clarify, the code used to update the market calculator HTML as the user makes selections, the need for object id selectors for most of the new content was eliminated. Selections for each drop down menu or checkbox added to the page are determined using the jQuery :selected and :checked selectors along with the ID selector for the row.

Additionally, the variable radioSelection was replaced with radio_selection following the formatting of other variables in the JavaScript code.