Added functions to remove dependence on object id
To remove the extraneous object ids that were inconsistently named and confusing relative to the ids used for the encapsulating row tags, they've been removed entirely. In earlier versions of calculator.js, the ids for the objects were needed to be able to place the drop down or button group contents in the correct place in the DOM. To get around this requirement, two new functions were defined, one each for buttons and drop downs, that combine the insertion and population of the objects into a single function. Inside these new functions, a temporary id is created and then removed after it is used by the populate function.

A handful of small, unrelated edits were made throughout to correct small mistakes. Closes #55.