Added tests for "walk" and "list_generator" functions
The “walk_techdata” and “list_generator_techdata” functions in
“mseg_techdata.py” are now both tested.

The “walk_techdata” function is used to run through all the levels of
the input microsegments JSON (imported as a dict) and update the values
at each technology-level leaf node in the dict to the appropriate
performance, cost, and lifetime information for that technology.  The
“list_generator_techdata” function is used to find the appropriate
information to update each node with in “walk_techdata”.

The test of “walk_techdata” uses a simpler version of this function,
which runs through a sample input dict and updates its terminal leaf
node values with the chain of dict keys that lead to each node.

The test of “list_generator_techdata” checks that it generates correct
performance, cost, and lifetime information for a series of
technologies given the microsegments dict key chain for each technology
along with supporting performance/cost/lifetime data available from EIA
or internally defined by BTO.  The test also checks that when invalid
technology dict key chains are supplied to the
“list_generator_techdata” function, the function yields an error.

A few adjustments have been made to the “mseg_techdata.py” code itself
to ensure that the new tests pass.  One change of particular importance
is in the handling of the lists of matched rows for a technology in
“list_generator_techdata”.  Once finalized, these lists are now
converted into numpy arrays with variable names assigned to each of the
columns.  This allows variables (tech. performance, cost, etc.) to be
accessed by their names in later operations.  The previous commit had
not made this conversion, which led to errors in these operations.

Other adjustments include minor syntax corrections and organizational
changes as well as further clarification of some comments.