Fix 'other' end use calculator queries
The technology types under the 'other' end use in the market calculator were triggering silent JavaScript errors because the queries to the JSON file were invalid. The microsegments JSON file refers to the 'other' end use as 'other (grid electric)', which was not correctly reflected in the variable res_end_use_values. Resolving the problem required updating that variable and the few locations in the code where selected_end_use is checked to see if it is equal to 'other' (now 'other (grid electric)'). Closes #65.