Minor syntax adjustments and added test cases
A few minor syntax corrections were required before a full run of the
“mseg_techdata.py” routine could be completed.  Many of the corrections
are to the filtering names used to find technology performance, cost,
and lifetime data in EIA .txt files.

In addition, new microsegment input cases have been added to the
“list_generator_techdata” test.  One new case concerns the natural gas
heat pump technology, for which there is only a “typical” performance
and cost level in the associated EIA file (data on most of the other
technologies in this file are for multiple performance/cost levels).  A
few more tests concern cases where a technology being searched is not
discovered in supporting performance/cost/lifetime data arrays, which
should yield a ValueError.

A final note: rows in the EIA performance/cost input array for
non-lighting technologies are no longer removed from the array upon
being matched with an input microsegment technology.  The associated
test of this functionality has accordingly been removed.