Adjust handling of baseline and measure technology lifetimes
It was initially assumed that the lifetime for both an efficient
measure and its comparable baseline technology would be specified as
point values in the measures JSON.  Subsequently, we decided to import
the baseline technology lifetime information from a separate
base_costperflife JSON that would also include performance and cost
data for the given technology.  In this JSON, performance, cost, and
lifetime data are all broken out by year.

This commit adjusts “run.py” and associated tests to accommodate the
changed handling of baseline technology lifetime information.
Specifically, it adds a “lifetime” key to each measure’s master
microsegment; the year-by-year lifetime values associated with this key
are each averages across all baseline technologies that contribute to
the master microsegment.

The commit also adjusts the handling of the measure lifetime input,
which can now be assigned a probability distribution and specified
separately for multiple applicable microsegments (just like the measure
performance and cost inputs).  Note, however, that tests do not yet
fully address the new handling of measure technology lifetime,
particularly in the calculation of measure prioritization metrics; such
tests will be added in a subsequent commit.