Disable commercial markets in calculator
Since the underlying data to support commercial markets for the market calculator are not yet available, the commercial market selection is now disabled to reduce confusion and the anticipated completion date is indicated above the selection field. The reference to the commercial markets in the instruction text above the selection fields is struck through as well.