Added tests of measure cost, performance, and/or lifetime distribution cases
A distribution of measure lifetimes is now finally accommodated in the
calculation of measure prioritization metrics (“calc_metric_update”) in
“run.py”.  Tests have also been added to “run_test.py” to ensure that
the correct lists of output prioritization metrics are yielded in cases
where a user has assigned a probability distribution to measure cost,
performance, lifetime, or a combination of these three inputs.

A key part of the strategy for handling distributions on cost,
performance, and lifetime variables in “calc_metric_update” is to
ensure consistency in the length of these variables before they are
entered into the actual prioritization metric calculations.

For example, if a user sets point values for measure cost and
performance but a distribution of values for measure lifetime, the cost
and performance point values can be converted to arrays of the same
length as the lifetime variable, with the point values repeated across
the length of the array.  It is then straightforward to iterate across
these arrays together in the calculation of measure prioritization
metrics.