Added a few lines to support introduction of staging
This minor set of changes supports the future introduction of a staging
sub-routine in “run.py”.  In particular, each measure now has a
user-defined “prioritization metric” for use in staging it against
other measures.  “PrioritizationMetricsTest” has been updated in
“run_test.py” to accommodate this change.

A “staging” indicator has also been added to the beginning of “run.py”,
which determines whether a new “run_staging” method is executed on
active measures in the main portion of the code; this new method is
currently passed over, but will be filled out in future commits that
build up the execution of the measure staging procedure.