Roll back staging additions and clean up test code
The previous commit had added some lines to support the future
introduction of a measure staging routine.  As we have decided to
pursue a different method for identifying/removing overlapping savings
across multiple measures, most of these previous changes have been
rolled back in this commit.

This commit also adds handling of cases where energy/carbon savings
used to calculate economic metrics are zero, and adds a ‘CommonMethods’
class to the beginning of ‘run_test.py’.  This class provides a common
set of methods that are accessed by the tests that follow, helping to
condense the length of the test code.