Add consumer choice data to baseline microsegments
Baseline microsegments are now described in terms of consumer choice
modeling parameters in addition to existing cost, performance, and
lifetime parameters.

Consumer choice modeling parameters will be used to determine the
realistic market share that a candidate efficiency measure could
capture if it were to compete for adoption with the baseline
microsegment technology and other candidate efficiency measures.
Currently, we plan to calculate this measure market share in a two-step
process:

1) Using a bass diffusion model, determine which portion of
new/replacement purchases for a baseline microsegment will stay with
the baseline technology, and which will go to an efficient alternative
(without specifying which one).  Currently, the COBAM modeling
framework is expected to yield bass diffusion coefficient information
for the commercial building sector; it is not clear yet how this
information will be derived for the residential sector.

2) Using a microsegment-specific consumer choice model that weighs a
candidate technology’s capital and operating costs, determine how the
“efficient” market calculated in 1) is apportioned across candidate
efficiency measures.  Note that EIA consumer choice model data are
available for most microsegments.*

The above two steps may be combined in a future modeling improvement,
but it is proposed that all relevant data be stored in the “consumer
choice” branch that this commit establishes for each microsegment in
the JSON output.

Associated changes to mseg_techdata_test.py ensure that the
mseg_techdata.py routine correctly updates consumer choice information
for each baseline microsegment.  A few additional changes to the tests
were also needed to accommodate the recently revised structure of the
microsegments JSON input (e.g., the addition of leaf nodes with square
footage information; changes to the position of solar/electric water
heating in the JSON hierarchy).

Finally, the commit makes a small tweak in the code syntax (changes
single quotes to double quotes around strings).

* For the residential sector, EIA consumer choice data are available
for all “major” equipment microsegments (heating, cooling, water
heating, washing, drying, cooking, refrigeration, lighting).  With the
exception of lighting technologies, consumer choice data are broken
down by census division, building type, and equipment class; for
lighting technologies, the data apply across all categories.  Data are
not currently available for MELs or for window/envelope technologies;
this issue will be addressed in a future commit.