Add check for inconsistent refrigerator/freezer sub-type year bins
In the EIA residential technology data, refrigerator and freezer
technologies have sub-types (e.g. top, side, and bottom freezer
refrigerator, chest and upright freezer).  Performance and cost data
for these technologies are averaged for a given projection year bin
(e.g., 2009-2013) across all the technology sub-types.

This commit ensures that when the above averages are calculated, all
sub-types of the technology are represented.  For example, in the
refrigerator case, there are 3 technology sub-types, and thus for an
example 2009-2013 projection bin, there should be 3 performance values
and 3 cost values to average across.  Any greater or fewer than 3 of
these values will now throw a ValueError.  The fail test in FillYrsTest
has been updated to ensure that the error is indeed thrown in this
case.