Add market entry and/or exit years for a measure
Certain measures will only be available on the market after a certain
year, or will be phased out over time.  This commit ensures that there
are no new adoptions of a measure in a year when it is unavailable on
the market.  This functionality is achieved through the definition of
‘market_entry_year’ and ‘market_exit_year’ fields in the measures JSON.

The market entry/exit feature is tested in the
‘FindPartitionMasterMicrosegmentTest’ of ‘run_test.py’ through the
addition of two sample measures (#11 and #12) that specify market entry
or exit years for the sample measure.  Note that market entry and exit
years of ‘None’ result in the measure being run and adopted across all
years in the modeling time horizon.