Calculate total and added annual energy and carbon savings
This commit implements calculation of two types of energy and carbon
savings for a measure in a given year: total and added savings. While
total savings reflect the impact of all measure adoptions simulated up
until and including that year in the modeling time horizon, added
energy/carbon savings reflect the impact of only those measure
adoptions occurring in the given year of interest.

For example, consider two years in a modeling time horizon in which the
total baseline energy market for a measure remains at 1000 TBTUs across
both years, the competed energy market for a measure is 100 TBTUs
across both years, and new measure adoptions save 10 TBTUs of that
competed market in both years, yielding an efficient market of 990
TBTUs in year one and 980 TBTUs in year two (assuming measure lifetime
is greater than two years).  In year two, the total savings would thus
be 20 TBTUs (1000 TBTUs – 980 TBTUs), and the added savings would be 10
TBTUs (20 TBTUs – 10 TBTUs).

Generally, it is expected that the annual total energy/carbon savings
will be of greatest interest as an output, though it may be valuable to
show how added savings for each year in the modeling time horizon
contribute to the total savings for those years.  For now, the energy
and carbon savings outputs of the ‘calc_metric_update’ function in
‘run.py’ are assigned ‘(total)’ and ‘(added)’ tags.  The
‘PrioritizationMetricsTest’ in ‘run_test.py’ has been updated to check
that these two types of energy and carbon savings outputs are being
generated correctly.  A future commit will enable calculation of total
energy and carbon cost savings as well (currently only added cost
savings are calculated).