Reorganize master mseg, calculate total and added cost savings
This commit makes the following changes to ‘run.py’:

1) Reorganize cost information in the master mseg output of
‘mseg_find_partiton’ to be more streamlined and include new ‘total’,
‘competed’, and ‘efficient’ cost breakdowns for the stock, energy, and
carbon cost categories.  Note that for a given year of interest, total
cost is that associated with the baseline stock; competed cost is that
associated with the portion of baseline stock that is new or up for
replacement; and efficient cost is that associated with the baseline
stock given a certain degree of measure adoption.

2) Use the updated cost information to calculate both total and added
annual stock, energy, and carbon cost savings in ‘calc_metric_update’.*
 For a given year of interest, total annual cost savings for a measure
reflect the cost impacts of all measure adoptions up until and
including that year; added annual cost savings reflect the cost impacts
of only those measure adoptions occurring in the given year of interest.

Tests in ‘run_test.py’ remain fundamentally the same, but outputs are
updated to reflect the reorganized cost information and to test the
correct generation of total and added annual cost savings outputs.

* Typically, stock cost ‘savings’ (e.g., incremental capital cost) will
be negative since the measure cost is often higher than the cost of the
baseline technology.