Summarize key measure outputs and write to CSV
In order to enable testing of a full ‘run.py’ execution, a
‘measures_test’ JSON input is created, and a new ‘write_outputs’
function is added that takes a dict of key measure outputs and writes
them to a CSV file.

For now, the ‘measures_test’ JSON includes one sample measure,
concerning Energy Star CFL reflectors.  More sample measures will be
added subsequently.

Output data in CSV format will be plotted in R.  Going forward, we will
likely need to change the data post-processing approach to accommodate
the development of a web-based plotting interface.

Other slight changes in this commit include the renaming of ‘life_meas’
to ‘product_lifetime’ in the ‘calc_metric_update’ function of ‘run.py’
as well as in the ‘run_test.py’ routine, to be consistent with the
naming of the measure lifetime attribute in ’measures_test’ JSON.  A
new measure attribute now also indicates whether a measure is active
for the current analysis run.