Add 10 sample measures and restrict certain output calculations
This commit adds 10 more sample measures to the ’measures_test’ JSON
that span multiple emerging technologies R&D areas (HVAC, windows &
envelope, water heating, lighting, appliances).  Note that for the
windows & envelope measures, no baseline technology data is currently
in place in the ‘basecostperflife’ JSON.  For these measures,
therefore, relative savings percentages are used for performance
calculations (for now).*

To speed up the calculations a little, an if statement is added to the
‘calc_metric_update’ function that checks whether the current measure
is being competed against other measures.  If the measure is not being
competed, cash flow calculations and associated economic output metrics
like irr, payback, and cost of conserved energy/carbon are not
calculated.  The ‘PrioritizationMetricsTest’ in ‘run_test.py’ has been
updated to reflect this change.

* Note that since the current sample measures run is for an uncompeted
technical potential, baseline windows and envelope cost/lifetime
information is not needed.