Add Annualized Net Present Value output to economic calculations
This commit adds an Annualized Net Present Value (ANPV) output metric
to the ‘calc_metric_update’ function in ‘run.py’.  ANPV is calculated
in three different ways using capital cost, energy cost, and carbon
cost cashflows as inputs.  All cash flows are normalized to the
competed number of measure units for the given year and are incremental
relative to a baseline technology.

The ANPV of a measure’s capital and operating costs (e.g., energy, or
energy + carbon) will be used to calculate the ultimate share of the
market that a measure captures when competed against other measures
accessing the same market.  For the residential sector, the capital and
operating ANPVs are entered into a log-linear regression equation to
calculate the market shares; for the commercial sector, the ANPVs are
calculated under varying discount rate assumptions and then compared
across competing measures to determine market share.

The ‘PrioritizationMetricsTest’ and ‘MetricUpdateTest’ functions of
‘run_test.py’ have been updated to accommodate the new ANPV outputs.