Small bug fix (*NEED TO ADD TEST OF BUG)
When running a sample measure that applied to “all residential
refrigerators and freezers,” the ’mseg_find_partition’ function yielded
a measure master microsegment that included erroneously large values.
Incorrectly, the function had added a contributing microsegment to the
master microsegment that does not exist (a combination of the
‘refrigeration’ end use and ‘freezers’ technology type).

This commit makes a small correction to ‘mseg_find_partition’ to ensure
that such nonexistent contributing microsegments are not counted
towards a measure’s master microsegment. The change explicitly blanks
out a potential contributing microsegment if its dictionary key chain
is not found in the input dictionary from which all contributing
microsegment stock and energy data are drawn.

A sample ‘refrigerators/freezers’ measure should be added to
‘run_test.py’ in a future commit to ensure that this potential bug is
adequately accounted for.